\name{ilpkgs}
\alias{ilpkgs}

\title{ilpkgs}

\description{Function to calculate the square of a number.}

\usage{
ilpkgs(x)
}

\arguments{
  \item{x}
}

\value{The value returned is a vector. Its element/s is/are of class numeric.}

\author{Silvan Albin <silvan.albin@hotmail.com>}

\examples{
x <- c("ggplot2", "reshape", "foreign")
ilpkgs(x)
}

