% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important-package.R
\docType{package}
\name{important-package}
\alias{important}
\alias{important-package}
\title{important: Supervised Feature Selection}
\description{
Interfaces for choosing important predictors in supervised regression, classification, and censored regression models. Permuted importance scores (Biecek and Burzykowski (2021) \doi{10.1201/9780429027192}) can be computed for 'tidymodels' model fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://important.tidymodels.org/}
  \item \url{https://github.com/tidymodels/important}
  \item Report bugs at \url{https://github.com/tidymodels/important/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
