% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-predictor-best.R,
%   R/step-predictor-retain.R, R/step_predictor_desirability.R
\name{required_pkgs.step_predictor_desirability}
\alias{required_pkgs.step_predictor_desirability}
\alias{required_pkgs.important}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_predictor_desirability}(x, ...)

\method{required_pkgs}{step_predictor_desirability}(x, ...)

\method{required_pkgs}{step_predictor_desirability}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
