\name{square_foo}
\alias{square_foo}

\title{square_foo}

\description{Function to calculate the square of a number.}

\usage{
square_foo(x)
}

\arguments{
  \item{x}
}

\value{The value returned is a vector. Its element/s is/are of class numeric.}

\author{Silvan Albin <silvan.albin@hotmail.com>}

\examples{
square_foo(x)
}

