\name{squareFoo}
\alias{squareFoo}

\title{squareFoo}

\description{Function to calculate the square of a number.}

\usage{
squareFoo(x)
}

\arguments{
  \item{x}
}

\value{The value returned is a vector. Its element/s is/are of class numeric.}

\author{Silvan Albin <silvan.albin@hotmail.com>}

\examples{
squareFoo(x)
}

