% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-predictor-best.R
\name{step_predictor_best}
\alias{step_predictor_best}
\title{Supervised Feature Selection via Choosing the Top Predictors}
\usage{
step_predictor_best(
  recipe,
  ...,
  score,
  role = NA,
  trained = FALSE,
  prop_terms = 0.5,
  update_prop = TRUE,
  removals = NULL,
  skip = FALSE,
  id = rand_id("predictor_best")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[recipes:selections]{selections()}} for more details.}

\item{score}{The name of a single score function from the \pkg{filtro}
package, such as \code{\link[filtro:score_imp_rf]{filtro::score_imp_rf()}}, \code{\link[filtro:score_roc_auc]{filtro:: score_roc_auc()}}, etc.
See the Details and Examples sections below. This argument \emph{should be named}
when used.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{prop_terms}{The proportion of predictors that should be retained when
ordered by overall desirability. A value of \code{\link[hardhat:tune]{hardhat::tune()}} can also be
used.}

\item{update_prop}{A logical: should \code{prop_terms} be updated so that at least
one predictor will be retained?}

\item{removals}{A character string that contains the names of predictors that
should be removed. These values are not determined until \code{\link[recipes:prep]{recipes::prep()}}
is called.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[recipes:bake]{bake()}}? While all operations are baked when \code{\link[recipes:prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{step_predictor_best()} creates a \emph{specification} of a recipe step that will
perform feature selection by ...
}
\details{
This step ...

This step can potentially remove columns from the data set. This may
cause issues for subsequent steps in your recipe if the missing columns are
specifically referenced by name. To avoid this, see the advice in the
\emph{Tips for saving recipes and filtering columns} section of
\link[recipes:selections]{recipes::selections}.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble::tibble is
returned with columns \code{terms} and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected to be removed}
\item{id}{character, id of this step}
}
}

