% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.importance_perm}
\alias{autoplot.importance_perm}
\title{Visualize importance scores}
\usage{
\method{autoplot}{importance_perm}(
  object,
  top = Inf,
  metric = NULL,
  eval_time = NULL,
  type = "importance",
  ...
)
}
\arguments{
\item{object}{A tibble of results from \code{\link[=importance_perm]{importance_perm()}}.}

\item{top}{An integer for how many terms to show. The rankings of predictors
are computed across metrics.}

\item{metric}{A character vector or \code{NULL} for which metric to plot. By
default, all metrics will be shown via facets. Possible options are
the entries in \code{.metric} column of the object.}

\item{eval_time}{For censored regression models, a vector of time points at
which the survival probability is estimated.}

\item{type}{A character value. The default is \code{"importance"} which shows the
overall signal-to-noise ration (i.e., mean divided by standard error).
Alternatively, \code{"direction"} shows the mean difference value with standard
error bounds.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualize importance scores
}
