% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important-package.R
\docType{package}
\name{important-package}
\alias{important}
\alias{important-package}
\title{important: Tools for Supervised Feature Selection}
\description{
Low-level data and interfaces to choose important predictors for predicting an outcome.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software PBC [copyright holder]
}

}
\keyword{internal}
