% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_predictor_desirability.R
\name{step_predictor_desirability}
\alias{step_predictor_desirability}
\title{Feature Selection}
\usage{
step_predictor_desirability(
  recipe,
  ...,
  score,
  role = NA,
  trained = FALSE,
  prop_terms = 0.5,
  update_prop = TRUE,
  results = NULL,
  removals = NULL,
  skip = FALSE,
  id = rand_id("predictor_desirability")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[recipes:selections]{selections()}} for more details.}

\item{score}{An object produced by \code{\link[desirability2:desirability]{desirability2::desirability()}} that uses
one or more score functions from the \pkg{filtro} package. See the Details
and Examples sections below. This argument \emph{should be named} when used.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{prop_terms}{The proportion of predictors that should be retained when
ordered by overall desirability.}

\item{update_prop}{A logical: should \code{prop_terms} be updated so that at least
one predictor will be retained?}

\item{results}{A data frame of score and desirability values for each
predictor evaluated. These values are not determined until
\code{\link[recipes:prep]{recipes::prep()}} is called.}

\item{removals}{A character string that contains the names of predictors that
should be removed. These values are not determined until \code{\link[recipes:prep]{recipes::prep()}}
is called.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[recipes:bake]{bake()}}? While all operations are baked when \code{\link[recipes:prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_predictor_desirability()} creates a \emph{specification} of a recipe step
that uses one or more "score" functions to measure how how much each
predictor is related to the outcome value. These scores are combined into
a composite value using user-specified \emph{desirability} functions and a
proportion of the most desirable predictors are retained.
}
\details{
This step ...

This step can potentially remove columns from the data set. This may
cause issues for subsequent steps in your recipe if the missing columns are
specifically referenced by name. To avoid this, see the advice in the
\emph{Tips for saving recipes and filtering columns} section of
\link[recipes:selections]{recipes::selections}.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble::tibble is
returned with columns \code{terms} and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected to be removed}
\item{id}{character, id of this step}
}
}

\section{Tuning Parameters}{
This step has 1 tuning parameters:
\itemize{
\item \code{prop_terms}: Threshold (type: double, default: 0.5)
}

The underlying operation does not allow for case weights.
}

\examples{
library(recipes)
library(desirability2)

if (rlang::is_installed("modeldata")) {
	# The `ad_data` has a binary outcome column ("Class") and mostly numeric
	# predictors. For these, we score the predictors using an analysis of
	# variance model where the predicor is the outcome and the outcome class
	# defines the groups.
	# There is also a single factor predictor ("Genotype") and we'll use
	# Fisher's Exact test to score it. NOTE that for scores using hypothesis
	# tests, the -log10(pvalue) is returned so that larger values are more
	# important.

	# The score_* objects here are from the filtro package. See Details above.
	goals <-
		desirability(
			maximize(score_xtab_pval_fisher),
			maximize(score_aov_pval)
		)

 example_data <- modeldata::ad_data
	rec <-
		recipe(Class ~ ., data = example_data) |>
		step_predictor_desirability(
			all_predictors(),
			score = goals,
			prop_terms = 1/2
		)
	rec

	# Now evaluate the predictors and rank them via desirability:
	rec_trained <- prep(rec)
	rec_trained

	# Use the tidy() method to get the results:
	predictor_scores <- tidy(rec_trained, number = 1)
	mean(predictor_scores$retained)
	predictor_scores

 # --------------------------------------------------------------------------

# Case-weight example: use the hardhat package to create the appropriate type
# of case weights. Here, we'll increase the weights for the minority class and
# add them to the data frame.

library(hardhat)

example_weights <- example_data
weights <- ifelse(example_data$Class == "Impaired", 5, 1)
example_weights$weights <- importance_weights(weights)

# To see if the scores can use case weights, load the filtro package and
# check the `case_weights` property:

library(filtro)

score_xtab_pval_fisher@case_weights
score_aov_pval@case_weights

# The recipe will automatically find the case weights and will
# not treat them as predictors.
rec_wts <-
	recipe(Class ~ ., data = example_weights) |>
	step_predictor_desirability(
		all_predictors(),
		score = goals,
		prop_terms = 1/2
	) |>
	prep()
rec_wts

predictor_scores_wts <-
	tidy(rec_wts, number = 1) |>
	select(terms, .d_overall_weighted = .d_overall)

library(dplyr)
library(ggplot2)

# The selection did not substantually change with these case weights
full_join(predictor_scores, predictor_scores_wts, by = "terms") |>
	ggplot(aes(.d_overall, .d_overall_weighted)) +
	geom_abline(col = "darkgreen", lty = 2) +
	geom_point(alpha = 1 / 2) +
	coord_fixed(ratio = 1) +
	labs(x = "Unweighted", y = "Class Weighted")
}
}
\references{
Derringer, G. and Suich, R. (1980), Simultaneous Optimization of
Several Response Variables. \emph{Journal of Quality Technology}, 12, 214-219.
}
\seealso{
\code{\link[desirability2:desirability]{desirability2::desirability()}}
}
